requirements.txt
asgiref==3.8.1
Django==5.1.2
django-taggit==6.1.0
pillow==11.0.0
python-decouple==3.8
sqlparse==0.5.1
